fix s22 audio exporting
this removes bink audio support, so old versions of the game's audio files will not work with this code

this will not be merged into the main branch, as this was solely to get a working version of the Rad Audio decoder so that it can be implemented in RSX once audio banks are supported.
the decoder outputs a buffer of floats instead of the previously exported 16-bit PCM data from bink audio. while it is possible to implement conversion for this data to 16-bit PCM, there is no real point for demonstration code. in RSX there will be an option to convert the audio to 16-bit PCM or leave it as 32-bit floats
the audio lib and headers are not provided as they are licensed under the Unreal Engine License, which prohibits the redistribution of source code to non-licensed individuals